+++ utils/module_browser/bootstrap/__init__.py
@@ -1,0 +1,18 @@
+"""
+Module: utils/module_browser/bootstrap/__init__.py
+Author: AlexZ1967
+Last updated: 2026-02-13
+
+Description:
+    Repository bootstrap helpers.
+"""
+
+from .repo_bootstrap_ops import (
+    bootstrap_module_remote_from_manager,
+    comfyui_requirements_path,
+)
+
+__all__ = [
+    "comfyui_requirements_path",
+    "bootstrap_module_remote_from_manager",
+]